test: split indirect eval import tests
Split indirect eval import tests as they depends on the JS stack to
resolve the referrer.
